Added PWA and it now works locally!
Progressive Web Apps allow a website to be installed on a mobile device
as if it were a normal app. Unfortunately, this is only possible with
websites that are being served from 127.0.0.1 OR when they are served
using HTTPS. For now, I've tested it using EDGE (because firefox does
not support it) and indeed I can install the PWA on my desktop computer
and it shows up as a normal app (without the location bar).